update navigation; recordgroupslistpage -> projectpage
